*: use the term 'Free Software'
While I've traditionally not agreed with the zealotry that the FSF
attacks proprietary companies, I do agree with Richard Stallman that the
term "open source" does focus on the wrong issues. Free software is
important because of the freedom it provides, not the quality of code it
produces. N.B. There are certain points (specifically with regards to
"Tivoisation"), where I very strongly disagree with the FSF.