doc: published: update 'Licenses, licenses, licenses'
I've recently changed my opinion on the GPLv3, mainly due to
conversations I've had with Richard Stallman. While I always agreed with
the philosophical issue, I felt that it wasn't the place of a software
license to impose restrictions on distributors. But then I realised that
it's that distributors decision to use free software in the first place,
so they should play ball with the ethical foundations of the software
they wish to use.