*: use GNU/Linux
GNU/Linux is the correct name of the combination of the (much larger)
GNU system and the (important) Linux system. Given Torvalds' lack of
interest in user freedom and pushing of the term "open source", it is
not fair to subvert the values of the GNU system in order to save 4
letters.